pyproject.toml
[tool.poetry]
name = "listenbravo"
version = "0.1.0"
description = ""
authors = ["Aditya K."]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.22.0"
pandas = "^2.0.0"
sqlalchemy = "^2.0.0"
langchain = "^0.1.0"
python-dotenv = "^1.0.0"
google-generativeai = "^0.7.2"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"